pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "cluster_score"
requires-python = ">= 3.10"
version = "0.0.1"
dependencies = [
    "pydicom",
    "numpy",
    "pandas",
    "pyyaml",
    "argparse",
    "scikit-image",
    "openpyxl"
]
